Autosaving knockout bindings, and restful urls for people, pets, and organizations.
With this update, we can now apply a binding to an element which will cause it to "autosave".  This means that shortly after the value is changed, an ajax request is fired off to save the field.  We're saving fields at a time
so that we can later track activity against every field changed (and by what user).