Combine duplicate events for pets into one single event.
When a user updates multiple things for a pet in sucession, multiple events will be added for the pet.  This causes the history log to get spammed with every single minor update for the pet.  Instead, lets combine these events into one event IF no other change has been made to the pet since.