Use UserService as spring security's UserDetailsService.
There appears to be an issue with the default UserDetailsService shipped with spring-security-core plugin when using Gorm.  In production, I'm getting a "can't use criteria" error.  By using our own UserService, which is known to correctly find users, we should get around this issue.

As an added benefit, user lookups can work within our application caching as well, so no database calls need to be made for unchanged users.