Add logic for resetting account passwords.
This update includes the controller and service support required in order to
reset user passwords.  Since admin accounts are setup without a known password,
this is the minimum that's required in order to update the accounts to a known
state.

Note that this does not include the views so that regular users can reset their
passwords.